api/types: move Port to api/types/container
This moves the `Port` type to the container package, and
deprecates the old location.
