+++ api/types/container/port.go
@@ -1,4 +1,4 @@
-package types
+package container
 
 // This file was generated by the swagger tool.
 // Editing this file might prove futile when you re-run the swagger generate command
