+++ hack/generate-swagger-api.sh
@@ -10,8 +10,7 @@ swagger generate model -f api/swagger.yaml \
 	-n PluginDevice \
 	-n PluginMount \
 	-n PluginEnv \
-	-n PluginInterfaceType \
-	-n Port
+	-n PluginInterfaceType
 
 swagger generate model -f api/swagger.yaml \
 	-t api -m types/container --skip-validator -C api/swagger-gen.yaml \
@@ -19,7 +18,8 @@ swagger generate model -f api/swagger.yaml \
 	-n ContainerWaitResponse \
 	-n ContainerWaitExitError \
 	-n ChangeType \
-	-n FilesystemChange
+	-n FilesystemChange \
+	-n Port
 
 swagger generate model -f api/swagger.yaml \
 	-t api -m types/image --skip-validator -C api/swagger-gen.yaml \
